middleware: Electrum Tunnel Unite tests
This commit adapts the unit tests to the changes introduced in the
previous commit for the electrum tunneling. The electrum client is now
only spawned with the first electrum rpc call.

The electrum client opens a new tcp connection to the electrum server
for each connected client of the middleware.